+++ plugins/calibre-bridge/__init__.py
@@ -9,7 +9,7 @@ class BinderyBridge(InterfaceActionBase):
     )
     supported_platforms = ['windows', 'osx', 'linux']
     author = 'vavallee'
-    version = (0, 1, 0)
+    version = (0, 3, 0)
     minimum_calibre_version = (6, 0, 0)
 
     actual_plugin = 'calibre_plugins.bindery_bridge.plugin:BinderyBridgeAction'